correlation-layer/requirements.txt
kafka-python
pyyaml